backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
sqlalchemy
pillow
pymupdf
python-multipart
